Se termina modulos secundarios, falta mejorar la funcionalidad de reporte hora

+++ SisteCredito-PruebaTecnica-FrontEnd/src/app/cms/pages/employee/employee.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { EmpleadosService } from 'src/app/services/empleados/empleados.service';
 import { Router } from '@angular/router';
 import { Empleados } from 'src/app/Models/empleados';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-employee',
@@ -41,12 +42,29 @@ export class EmployeeComponent implements OnInit {
   }
 
   deleteEmployee(id:string){
-    this.empleadosService.deleteEmployee(id)
-    .subscribe({next: (response) => {
-      this.listEmployees();
-    }, error: (response) => {
-      console.log(response)
-    }
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Si eliminas el registro sera de forma permanente",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.empleadosService.deleteEmployee(id)
+        .subscribe({next: (response) => {
+          this.listEmployees();
+        }, error: (response) => {
+          console.log(response)
+        }
+        })
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
     })
   }
